Move sync logic from on-heap index to aggregator

+++ processing/src/main/java/io/druid/query/aggregation/HistogramAggregator.java
@@ -52,7 +52,7 @@ public class HistogramAggregator implements Aggregator
   @Override
   public void aggregate()
   {
-    histogram.offer((float)selector.get());
+    histogram.offer(selector.get());
   }
 
   @Override
